Remove link to same page
This change removes the link to README.md within README.md.

Personally, when reading the Table of Contents section, I tend to click the Introduction link, thinking it's a new page. When that leads back to the same page, I get a bit confused, until I check the link and see that it points to the page I'm on. This change helps to clarify that the current page is included in the TOC.